Create modules for main.rb
